requirements.txt
numpy==1.26.4
mypy==1.4.1
pillow==10.2.0
scipy==1.12.0
torch==2.2.1
torchvision==0.17.1
#-e git+https://codeberg.org/alexander-g/traininglib@7453971aba0bd40dab0384ab487d391b0a49beca#egg=traininglib